adding nomor 4 5

+++ src/main/java/com/apap/tugas/repository/PerpustakaanDb.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PerpustakaanDb extends JpaRepository<PerpustakaanModel, Long> {
-	PerpustakaanModel findById(long id);
+public interface PerpustakaanDb extends JpaRepository<PerpustakaanModel, Integer> {
+	PerpustakaanModel findById(int id);
 }
