+++ src/main/java/com/apap/tugas/service/PerpustakaanService.java
@@ -3,5 +3,5 @@ package com.apap.tugas.service;
 import com.apap.tugas.model.PerpustakaanModel;
 
 public interface PerpustakaanService {
-	PerpustakaanModel getPerpustakaanById(long Id);
+	PerpustakaanModel getPerpustakaanById(int Id);
 }
